+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 
+### Added
+- resource `pullzone`: vary cache by request headers;
+
 ### Changed
 - resource `pullzone`: `websockets_max_connections` can be configured to increments of 100;